requirements.txt
functions-framework==3.*
google-genai
python-dotenv